Add files via upload
Step 1: Read in the Data
Read the data into R
List the structure of the data (str)
Execute a summary of the data
Print the first six records

Step 2: Box-Whisker Plots
Plot a box plot of all the numeric variables split by the grouping variable. The plot needs the following:
Add color to the boxes.
Comment on whether or not there are any observable differences in the box plots between the two groups.

Step 3: Histograms
Plot a histogram of at least one of the numeric variables.
Manually set the number of breaks to a value that makes sense.
Superimpose a density line to the graph.

Step 4: Impute "Fix" all the numeric variables that have missing values.
For the missing Target variables, simply set the missing values to zero.
For the remaining numeric variables with missing values, create two new variables. One variable will have a name beginning with IMP_ and it will contained the imputed value. The second value will have a name beginning with M_ and it will contain a 1 if the record was imputed and a zero if it was not.
The median or mean value will be useful in most cases.
Delete the original variable after it has been imputed.
Run a summary to prove that all the variables have been imputed.
Compute a sum for all the M_ variables to prove that the number of flags is equal to the number of missing values.

Step 5: One Hot Encoding
For the character / category variables, perform one hot encoding. For this create a Flag for each categories.
Delete the original class variable
Run a summary to show that the category variables have been replaced by Flag variables.